requirements.txt
adapt==0.3.1
common==0.1.2
dalib==0.2
dataloader==2.0
matplotlib==3.1.3
mnistusps==0.1.0
nltk==3.4.5
numpy==1.21.6
office31==0.1.3
pandas==1.3.5
Pillow==9.2.0
scikit_learn==1.1.2
tabulate==0.8.7
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
wandb
timm
